+++ packages/experimental/Menu/src/MenuList/MenuList.tsx
@@ -4,6 +4,9 @@ import { View } from 'react-native';
 import { compose, UseSlots, withSlots } from '@fluentui-react-native/framework';
 import { menuListName, MenuListProps, MenuListType } from './MenuList.types';
 import { stylingSettings } from './MenuList.styling';
+import { MenuListProvider } from '../context/menuListContext';
+import { useMenuList } from './useMenuList';
+import { useMenuListContextValue } from './useMenuContextValue';
 
 export const MenuList = compose<MenuListType>({
   displayName: menuListName,
@@ -12,10 +15,16 @@ export const MenuList = compose<MenuListType>({
     root: View,
   },
   useRender: (userProps: MenuListProps, useSlots: UseSlots<MenuListType>) => {
-    const Slots = useSlots(userProps);
+    const menuList = useMenuList(userProps);
+    const contextValue = useMenuListContextValue(menuList);
+    const Slots = useSlots(menuList);
 
     return (_final: MenuListProps, children: React.ReactNode) => {
-      return <Slots.root>{children}</Slots.root>;
+      return (
+        <MenuListProvider value={contextValue}>
+          <Slots.root>{children}</Slots.root>
+        </MenuListProvider>
+      );
     };
   },
 });